Climb with Sensors
Added Sensor Commands and methods.
